requirements.txt
astor==0.8.1
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
cryptography==41.0.7
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pyOpenSSL==23.3.0
PySocks==1.7.1
pytest==7.4.3
python-dotenv==1.0.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
trustme==1.1.0
urllib3==2.1.0
wsproto==1.2.0
